+++ src/frontend/apps/impress/src/features/service-worker/__tests__/useSWRegister.test.tsx
@@ -11,7 +11,6 @@ const TestComponent = () => {
 
 describe('useSWRegister', () => {
   it('checks service-worker is register', () => {
-    process.env.NEXT_PUBLIC_BUILD_ID = '123456';
     vi.spyOn(console, 'error').mockImplementation(() => {});
 
     const registerSpy = vi.fn();
@@ -35,7 +34,7 @@ describe('useSWRegister', () => {
 
     render(<TestComponent />);
 
-    expect(registerSpy).toHaveBeenCalledWith('/service-worker.js?v=123456');
+    expect(registerSpy).toHaveBeenCalledWith('/service-worker.js');
     expect(addEventListenerSpy).toHaveBeenCalledWith(
       'controllerchange',
       expect.any(Function),
@@ -44,7 +43,6 @@ describe('useSWRegister', () => {
 
   it('checks service-worker is not register', () => {
     process.env.NEXT_PUBLIC_SW_DEACTIVATED = 'true';
-    process.env.NEXT_PUBLIC_BUILD_ID = '123456';
 
     const registerSpy = vi.fn();
     registerSpy.mockImplementation(
@@ -62,6 +60,6 @@ describe('useSWRegister', () => {
 
     render(<TestComponent />);
 
-    expect(registerSpy).not.toHaveBeenCalledWith('/service-worker.js?v=123456');
+    expect(registerSpy).not.toHaveBeenCalledWith('/service-worker.js');
   });
 });