+++ src/frontend/apps/impress/src/features/service-worker/hooks/useSWRegister.tsx
@@ -7,7 +7,7 @@ export const useSWRegister = () => {
       process.env.NEXT_PUBLIC_SW_DEACTIVATED !== 'true'
     ) {
       navigator.serviceWorker
-        .register(`/service-worker.js?v=${process.env.NEXT_PUBLIC_BUILD_ID}`)
+        .register(`/service-worker.js`)
         .then((registration) => {
           registration.onupdatefound = () => {
             const newWorker = registration.installing;
@@ -29,11 +29,13 @@ export const useSWRegister = () => {
           console.error('Service worker registration failed:', err);
         });
 
-      const currentController = navigator.serviceWorker.controller;
+      let refreshing = false;
       const onControllerChange = () => {
-        if (currentController) {
-          window.location.reload();
+        if (refreshing) {
+          return;
         }
+        refreshing = true;
+        window.location.reload();
       };
       navigator.serviceWorker.addEventListener(
         'controllerchange',